revert name changes
